test(combat-trainer): consolidate 8 spec files into one canonical spec
Merge combat_trainer_bug_fixes, safety, warhorn_egg, slivers, gempouch, force_cleanup, and almanac spec files into spec/combat_trainer_spec.rb so there is a single spec per script.

Unify the scaffolding that conflicted when the files were co-loaded: one load_lic_class and one RSpec.configure; reopen Harness::DRSpells instead of a top-level shadow that dropped active_spells; a single UserVars store covering every accessor used; the union of the DRC and DRCI command stubs; and load all ten combat-trainer classes.

No test intent changed. The merged file runs the same 260 examples with the same 22 pre-existing failures the eight files produced separately. rubocop clean.
