+++ spec/combat_trainer_spec.rb
@@ -69,6 +69,13 @@ module DRCI
     def inside?(*_args) = false
     def fill_gem_pouch_with_container(*_args) = nil
     def count_all_boxes(*_args) = 0
+    # -- union additions from merged specs --
+    def stow_item?(*_args) = true
+    def remove_item?(*_args) = true
+    def wearing?(*_args) = false
+    def swap_out_full_gempouch?(*_args) = true
+    def get_item_if_not_held?(*_args) = true
+    def exists?(*_args) = true
   end
 end
 
@@ -137,16 +144,57 @@ module Script
   def self.running?(*_args) = false
 end
 
+# Unified UserVars store. Backs every UserVars key the combat-trainer code
+# and the merged specs touch (moons, sun, discerns, friends, warhorn,
+# almanac_last_use, yiamura, paladin_last_badge_use, combat_trainer_debug,
+# ...) via a dynamic data hash. moons reads back a sane default so the
+# slivers specs see { 'visible' => [] } when unset, and _set_moons/_reset
+# match the helpers those specs call.
 class UserVars
+  @data = {}
+
   class << self
-    attr_accessor :moons unless method_defined?(:moons)
-    attr_accessor :sun unless method_defined?(:sun)
-    attr_accessor :discerns unless method_defined?(:discerns)
-    attr_accessor :friends unless method_defined?(:friends)
+    def _data
+      @data ||= {}
+    end
+
+    def _reset
+      @data = {}
+    end
+
+    def moons
+      _data[:moons] || { 'visible' => [] }
+    end
+
+    def moons=(val)
+      _data[:moons] = val
+    end
+
+    def _set_moons(val)
+      _data[:moons] = val
+    end
+
+    def method_missing(name, *args)
+      key = name.to_s
+      if key.end_with?('=')
+        _data[key.chomp('=').to_sym] = args.first
+      else
+        _data[name.to_sym]
+      end
+    end
+
+    def respond_to_missing?(_name, _include_private = false)
+      true
+    end
   end
 end
 
-class DRSpells
+# Reopen the harness DRSpells (do NOT shadow with a fresh top-level class --
+# a fresh class would lose active_spells/_set_active_spells that several specs
+# rely on). Add known_spells and slivers backed by their own class vars, and
+# extend _reset to clear them too. Code under test resolves DRSpells to
+# Harness::DRSpells via the include, so this is the single shared class.
+class Harness::DRSpells
   @@_known_spells = {}
   @@_slivers = false
 
@@ -154,6 +202,15 @@ class DRSpells
   def self._set_known_spells(val) = (@@_known_spells = val)
   def self.slivers = @@_slivers
   def self._set_slivers(val) = (@@_slivers = val)
+
+  class << self
+    alias_method :_orig_reset, :_reset
+    def _reset
+      _orig_reset
+      @@_known_spells = {}
+      @@_slivers = false
+    end
+  end
 end
 
 $HUNTING_BUDDY = nil
@@ -181,6 +238,10 @@ load_lic_class('combat-trainer.lic', 'SetupProcess')
 load_lic_class('combat-trainer.lic', 'ManipulateProcess')
 load_lic_class('combat-trainer.lic', 'AttackProcess')
 load_lic_class('combat-trainer.lic', 'AbilityProcess')
+load_lic_class('combat-trainer.lic', 'SafetyProcess')
+load_lic_class('combat-trainer.lic', 'SpellProcess')
+load_lic_class('combat-trainer.lic', 'PetProcess')
+load_lic_class('combat-trainer.lic', 'TrainerProcess')
 
 RSpec.configure do |config|
   config.before(:each) do
@@ -201,6 +262,43 @@ def ct_setup
   $COMBAT_TRAINER = double('CombatTrainer', stop: nil)
   $right_hand = nil
   $left_hand = nil
+end
+
+# -- Top-level helpers from the merged warhorn/egg spec --
+# These are used by the warhorn AbilityProcess describe block below.
+# Other describe blocks shadow build_game_state with their own scoped
+# definitions, so these top-level versions only apply where no scoped
+# version exists.
+def build_ability_process(**overrides)
+  instance = AbilityProcess.allocate
+  defaults = {
+    warhorn_nouns: [],
+    egg_count: 0,
+    warhorn_or_egg: [],
+    warhorn_items: [],
+    egg_ids: [],
+    item_cooldowns: {},
+    warhorn_cooldown: 1200
+  }
+  defaults.merge(overrides).each do |k, v|
+    instance.instance_variable_set(:"@#{k}", v)
+  end
+  instance
+end
+
+def build_game_state(**attrs)
+  defaults = {
+    currently_whirlwinding: false
+  }
+  state = double('GameState', defaults.merge(attrs))
+  allow(state).to receive(:sheath_whirlwind_offhand)
+  allow(state).to receive(:wield_whirlwind_offhand)
+  state
+end
+
+def stub_right_hand_with_id(id)
+  hand = OpenStruct.new(name: 'item', noun: 'item', id: id)
+  allow(GameObj).to receive(:right_hand).and_return(hand)
 end
 
 # ===================================================================
@@ -1570,5 +1668,2221 @@ RSpec.describe 'State mutation after cleanup' do
     expect(gs.done_cleaning_up?).to be true
 
     expect { gs.next_clean_up_step }.not_to raise_error
+  end
+end
+
+# ###################################################################
+# MERGED FROM spec/combat_trainer_bug_fixes_spec.rb
+# ###################################################################
+
+# ===========================================================================
+# SetupProcess#last_stance -- nil guard for Flags['last-stance']
+# ===========================================================================
+RSpec.describe SetupProcess do
+  def build_setup_process
+    SetupProcess.allocate
+  end
+
+  describe '#last_stance' do
+    context 'when Flags[last-stance] has not fired yet (nil)' do
+      it 'returns zeroed stance hash instead of raising NoMethodError' do
+        Flags['last-stance'] = nil
+        instance = build_setup_process
+
+        result = instance.send(:last_stance)
+
+        expect(result).to eq({ 'EVASION' => 0, 'PARRY' => 0, 'SHIELD' => 0, 'SPARE' => 0 })
+      end
+    end
+
+    context 'when Flags[last-stance] has a valid stance string' do
+      it 'parses the percentages correctly' do
+        Flags['last-stance'] = ['80% pointed stance 60% pointed stance 40% pointed stance 20']
+        instance = build_setup_process
+
+        result = instance.send(:last_stance)
+
+        expect(result).to eq({ 'EVASION' => 80, 'PARRY' => 60, 'SHIELD' => 40, 'SPARE' => 20 })
+      end
+    end
+
+    context 'when Flags[last-stance] has all zeros' do
+      it 'returns all zeros' do
+        Flags['last-stance'] = ['0% pointed stance 0% pointed stance 0% pointed stance 0']
+        instance = build_setup_process
+
+        result = instance.send(:last_stance)
+
+        expect(result).to eq({ 'EVASION' => 0, 'PARRY' => 0, 'SHIELD' => 0, 'SPARE' => 0 })
+      end
+    end
+  end
+end
+
+# ===========================================================================
+# ManipulateProcess#manipulate -- ordinal targeting for duplicate NPCs
+# ===========================================================================
+RSpec.describe ManipulateProcess do
+  def build_manipulate_process(**overrides)
+    instance = ManipulateProcess.allocate
+    defaults = {
+      threshold: 5,
+      manip_to_train: false,
+      last_manip: Time.now - 200,
+      filtered_npcs: []
+    }
+    defaults.merge(overrides).each do |k, v|
+      instance.instance_variable_set(:"@#{k}", v)
+    end
+    instance
+  end
+
+  def build_game_state(**attrs)
+    defaults = {
+      npcs: [],
+      danger: false,
+      construct_mode?: false
+    }
+    state = double('GameState', defaults.merge(attrs))
+    allow(state).to receive(:construct?).and_return(false)
+    allow(state).to receive(:construct)
+    state
+  end
+
+  describe '#manipulate' do
+    before(:each) do
+      allow(DRC).to receive(:bput).and_return('You attempt to empathically manipulate')
+    end
+
+    context 'when all NPCs have different nouns' do
+      it 'uses "first" ordinal for each NPC' do
+        game_state = build_game_state
+        instance = build_manipulate_process(
+          threshold: 3,
+          filtered_npcs: %w[rat kobold goblin]
+        )
+
+        instance.send(:manipulate, game_state)
+
+        expect(DRC).to have_received(:bput).with(/manipulate friendship first rat/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).to have_received(:bput).with(/manipulate friendship first kobold/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).to have_received(:bput).with(/manipulate friendship first goblin/, anything, anything, anything, anything, anything, anything)
+      end
+    end
+
+    context 'when multiple NPCs share the same noun' do
+      it 'uses incrementing ordinals for duplicate nouns' do
+        game_state = build_game_state
+        instance = build_manipulate_process(
+          threshold: 3,
+          filtered_npcs: %w[rat rat rat]
+        )
+
+        instance.send(:manipulate, game_state)
+
+        expect(DRC).to have_received(:bput).with(/manipulate friendship first rat/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).to have_received(:bput).with(/manipulate friendship second rat/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).to have_received(:bput).with(/manipulate friendship third rat/, anything, anything, anything, anything, anything, anything)
+      end
+    end
+
+    context 'when mixed duplicate and unique NPCs are present' do
+      it 'tracks ordinals independently per noun' do
+        game_state = build_game_state
+        instance = build_manipulate_process(
+          threshold: 4,
+          filtered_npcs: %w[rat kobold rat kobold]
+        )
+
+        instance.send(:manipulate, game_state)
+
+        expect(DRC).to have_received(:bput).with(/manipulate friendship first rat/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).to have_received(:bput).with(/manipulate friendship first kobold/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).to have_received(:bput).with(/manipulate friendship second rat/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).to have_received(:bput).with(/manipulate friendship second kobold/, anything, anything, anything, anything, anything, anything)
+      end
+    end
+
+    context 'when an NPC is a construct' do
+      it 'skips constructs and does not increment ordinal for that noun' do
+        game_state = build_game_state
+        allow(game_state).to receive(:construct?).with('golem').and_return(true)
+        allow(game_state).to receive(:construct?).with('rat').and_return(false)
+
+        instance = build_manipulate_process(
+          threshold: 2,
+          filtered_npcs: %w[golem rat]
+        )
+
+        instance.send(:manipulate, game_state)
+
+        expect(DRC).not_to have_received(:bput).with(/manipulate friendship .* golem/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).to have_received(:bput).with(/manipulate friendship first rat/, anything, anything, anything, anything, anything, anything)
+      end
+    end
+
+    context 'when threshold limits the number of manipulations' do
+      it 'stops after reaching the threshold' do
+        game_state = build_game_state
+        instance = build_manipulate_process(
+          threshold: 2,
+          filtered_npcs: %w[rat rat rat]
+        )
+
+        instance.send(:manipulate, game_state)
+
+        expect(DRC).to have_received(:bput).with(/manipulate friendship first rat/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).to have_received(:bput).with(/manipulate friendship second rat/, anything, anything, anything, anything, anything, anything)
+        expect(DRC).not_to have_received(:bput).with(/manipulate friendship third rat/, anything, anything, anything, anything, anything, anything)
+      end
+    end
+  end
+end
+
+# ###################################################################
+# MERGED FROM spec/combat_trainer_safety_spec.rb
+# ###################################################################
+
+RSpec.describe SafetyProcess do
+  # The original safety spec seeded $HUNTING_BUDDY/$COMBAT_TRAINER and reset
+  # known_spells in its own RSpec.configure before(:each). ct_setup is a
+  # superset of that, so we use it here to preserve the same per-example state.
+  before(:each) { ct_setup }
+
+  def build_safety_process(**overrides)
+    instance = SafetyProcess.allocate
+    defaults = {
+      equipment_manager: double('EquipmentManager'),
+      health_threshold: 20,
+      stop_on_bleeding: true,
+      safety_untendable_threshold: 3,
+      safety_exit_on_bleeding: false,
+      safety_concentration_minimum: nil,
+      safety_escape_health_threshold: nil,
+      untendable_counter: 0
+    }
+    defaults.merge(overrides).each do |k, v|
+      instance.instance_variable_set(:"@#{k}", v)
+    end
+    instance
+  end
+
+  def build_game_state(**attrs)
+    defaults = {
+      danger: false,
+      retreating?: false
+    }
+    state = double('GameState', defaults.merge(attrs))
+    allow(state).to receive(:danger=)
+    state
+  end
+
+  # Stub the rest of execute that runs after the safety branches
+  def stub_post_safety(instance)
+    allow(instance).to receive(:check_item_recovery)
+    allow(instance).to receive(:tend_lodged)
+    allow(instance).to receive(:tend_parasite)
+    allow(instance).to receive(:active_mitigation)
+    allow(instance).to receive(:in_danger?).and_return(false)
+    allow(instance).to receive(:keep_away)
+    allow(instance).to receive(:bleeding?).and_return(false)
+    allow(instance).to receive(:stunned?).and_return(false)
+    DRStats.health = 100
+    DRStats.concentration = 100
+    allow(DRCA).to receive(:activate_khri?).and_return(true)
+  end
+
+  describe '#execute' do
+    describe 'safety_untendable_threshold' do
+      it 'stops hunt at default threshold of 3' do
+        instance = build_safety_process(untendable_counter: 3)
+        stub_post_safety(instance)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).to have_received(:stop_hunting)
+        expect($COMBAT_TRAINER).to have_received(:stop)
+      end
+
+      it 'does not stop hunt below default threshold' do
+        instance = build_safety_process(untendable_counter: 2)
+        stub_post_safety(instance)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).not_to have_received(:stop_hunting)
+      end
+
+      it 'stops hunt at custom threshold of 1' do
+        instance = build_safety_process(safety_untendable_threshold: 1, untendable_counter: 1)
+        stub_post_safety(instance)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).to have_received(:stop_hunting)
+      end
+
+      it 'requires stop_on_bleeding to be true' do
+        instance = build_safety_process(untendable_counter: 3, stop_on_bleeding: false)
+        stub_post_safety(instance)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).not_to have_received(:stop_hunting)
+      end
+    end
+
+    describe 'safety_concentration_minimum' do
+      it 'stops hunt when concentration drops below minimum' do
+        instance = build_safety_process(safety_concentration_minimum: 10)
+        stub_post_safety(instance)
+        DRStats.concentration = 5
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).to have_received(:stop_hunting)
+      end
+
+      it 'does not stop hunt when concentration is above minimum' do
+        instance = build_safety_process(safety_concentration_minimum: 10)
+        stub_post_safety(instance)
+        DRStats.concentration = 50
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).not_to have_received(:stop_hunting)
+      end
+
+      it 'is disabled when nil' do
+        DRStats.concentration = 0
+        instance = build_safety_process(safety_concentration_minimum: nil)
+        stub_post_safety(instance)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).not_to have_received(:stop_hunting)
+      end
+    end
+
+    describe 'safety_escape_health_threshold (Thief Vanish)' do
+      before(:each) do
+        DRStats.guild = 'Thief'
+        DRSpells._set_known_spells({ 'Vanish' => true })
+      end
+
+      it 'activates Vanish and stops hunt when health is below threshold' do
+        instance = build_safety_process(safety_escape_health_threshold: 90)
+        stub_post_safety(instance)
+        DRStats.health = 85
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect(DRCA).to have_received(:activate_khri?).with(false, "Vanish")
+        expect($HUNTING_BUDDY).to have_received(:stop_hunting)
+      end
+
+      it 'activates Vanish when bleeding' do
+        DRStats.health = 100
+        instance = build_safety_process(safety_escape_health_threshold: 90)
+        stub_post_safety(instance)
+        allow(instance).to receive(:bleeding?).and_return(true)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect(DRCA).to have_received(:activate_khri?).with(false, "Vanish")
+      end
+
+      it 'does not fire for non-Thieves' do
+        instance = build_safety_process(safety_escape_health_threshold: 90)
+        stub_post_safety(instance)
+        DRStats.guild = 'Ranger'
+        DRStats.health = 50
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect(DRCA).not_to have_received(:activate_khri?)
+      end
+
+      it 'does not fire if Thief does not know Vanish' do
+        instance = build_safety_process(safety_escape_health_threshold: 90)
+        stub_post_safety(instance)
+        DRSpells._set_known_spells({})
+        DRStats.health = 50
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect(DRCA).not_to have_received(:activate_khri?)
+      end
+
+      it 'is disabled when nil' do
+        instance = build_safety_process(safety_escape_health_threshold: nil)
+        stub_post_safety(instance)
+        DRStats.health = 10
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect(DRCA).not_to have_received(:activate_khri?)
+      end
+    end
+
+    describe 'safety_exit_on_bleeding' do
+      it 'stops hunt when bleeding and setting is true' do
+        instance = build_safety_process(safety_exit_on_bleeding: true)
+        stub_post_safety(instance)
+        allow(instance).to receive(:bleeding?).and_return(true)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).to have_received(:stop_hunting)
+      end
+
+      it 'stops hunt when stunned with low health' do
+        instance = build_safety_process(safety_exit_on_bleeding: true)
+        stub_post_safety(instance)
+        DRStats.health = 70
+        allow(instance).to receive(:stunned?).and_return(true)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).to have_received(:stop_hunting)
+      end
+
+      it 'does not stop hunt when stunned with high health' do
+        instance = build_safety_process(safety_exit_on_bleeding: true)
+        stub_post_safety(instance)
+        DRStats.health = 95
+        allow(instance).to receive(:stunned?).and_return(true)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).not_to have_received(:stop_hunting)
+      end
+
+      it 'does not fire when setting is false' do
+        instance = build_safety_process(safety_exit_on_bleeding: false)
+        stub_post_safety(instance)
+        allow(instance).to receive(:bleeding?).and_return(true)
+        game_state = build_game_state
+
+        instance.execute(game_state)
+
+        expect($HUNTING_BUDDY).not_to have_received(:stop_hunting)
+      end
+    end
+  end
+end
+
+# ###################################################################
+# MERGED FROM spec/combat_trainer_warhorn_egg_spec.rb
+# ###################################################################
+
+# ===========================================================================
+# AbilityProcess warhorn/egg discovery and usage
+# ===========================================================================
+RSpec.describe AbilityProcess do
+  before(:each) do
+    allow(DRC).to receive(:bput).and_return('Roundtime')
+    allow(DRC).to receive(:message)
+  end
+
+  # ===========================================================================
+  # #discover_egg
+  # ===========================================================================
+  describe '#discover_egg' do
+    it 'records the game ID when egg is found' do
+      instance = build_ability_process
+      stub_right_hand_with_id('12345')
+      allow(DRCI).to receive(:get_item?).with('egg').and_return(true)
+      allow(DRCI).to receive(:stow_item?).and_return(true)
+
+      instance.send(:discover_egg, 'egg')
+
+      expect(instance.instance_variable_get(:@egg_ids)).to eq(['12345'])
+    end
+
+    it 'stows by game ID after discovery' do
+      instance = build_ability_process
+      stub_right_hand_with_id('12345')
+      allow(DRCI).to receive(:get_item?).with('egg').and_return(true)
+      allow(DRCI).to receive(:stow_item?).and_return(true)
+
+      instance.send(:discover_egg, 'egg')
+
+      expect(DRCI).to have_received(:stow_item?).with('#12345')
+    end
+
+    it 'warns and does not record when egg is not found' do
+      instance = build_ability_process
+      allow(DRCI).to receive(:get_item?).with('second egg').and_return(false)
+
+      instance.send(:discover_egg, 'second egg')
+
+      expect(instance.instance_variable_get(:@egg_ids)).to be_empty
+      expect(DRC).to have_received(:message).with(/Could not find 'second egg'/)
+    end
+  end
+
+  # ===========================================================================
+  # #discover_warhorn
+  # ===========================================================================
+  describe '#discover_warhorn' do
+    it 'records worn warhorn when remove succeeds' do
+      instance = build_ability_process
+      stub_right_hand_with_id('99')
+      allow(DRCI).to receive(:remove_item?).with('warhorn').and_return(true)
+      allow(DRCI).to receive(:wear_item?).and_return(true)
+
+      instance.send(:discover_warhorn, 'warhorn')
+
+      items = instance.instance_variable_get(:@warhorn_items)
+      expect(items).to eq([{ id: '99', worn: true }])
+    end
+
+    it 're-wears a worn warhorn after discovery' do
+      instance = build_ability_process
+      stub_right_hand_with_id('99')
+      allow(DRCI).to receive(:remove_item?).with('warhorn').and_return(true)
+      allow(DRCI).to receive(:wear_item?).and_return(true)
+
+      instance.send(:discover_warhorn, 'warhorn')
+
+      expect(DRCI).to have_received(:wear_item?).with('#99')
+    end
+
+    it 'records stowed warhorn when remove fails but get succeeds' do
+      instance = build_ability_process
+      stub_right_hand_with_id('50')
+      allow(DRCI).to receive(:remove_item?).with('horn').and_return(false)
+      allow(DRCI).to receive(:get_item?).with('horn').and_return(true)
+      allow(DRCI).to receive(:stow_item?).and_return(true)
+
+      instance.send(:discover_warhorn, 'horn')
+
+      items = instance.instance_variable_get(:@warhorn_items)
+      expect(items).to eq([{ id: '50', worn: false }])
+    end
+
+    it 'warns when warhorn is not found at all' do
+      instance = build_ability_process
+      allow(DRCI).to receive(:remove_item?).with('horn').and_return(false)
+      allow(DRCI).to receive(:get_item?).with('horn').and_return(false)
+
+      instance.send(:discover_warhorn, 'horn')
+
+      expect(instance.instance_variable_get(:@warhorn_items)).to be_empty
+      expect(DRC).to have_received(:message).with(/Could not find warhorn 'horn'/)
+    end
+  end
+
+  # ===========================================================================
+  # #set_warhorn_or_egg
+  # ===========================================================================
+  describe '#set_warhorn_or_egg' do
+    it 'builds rotation with both egg and warhorn when both are found' do
+      instance = build_ability_process(egg_count: 1, warhorn_nouns: ['warhorn'])
+      stub_right_hand_with_id('10')
+      allow(DRCI).to receive(:get_item?).and_return(true)
+      allow(DRCI).to receive(:stow_item?).and_return(true)
+      allow(DRCI).to receive(:remove_item?).and_return(true)
+      allow(DRCI).to receive(:wear_item?).and_return(true)
+
+      instance.send(:set_warhorn_or_egg)
+
+      expect(instance.instance_variable_get(:@warhorn_or_egg)).to eq(%w[egg warhorn])
+    end
+
+    it 'builds rotation with only egg when no warhorns configured' do
+      instance = build_ability_process(egg_count: 1, warhorn_nouns: [])
+      stub_right_hand_with_id('10')
+      allow(DRCI).to receive(:get_item?).and_return(true)
+      allow(DRCI).to receive(:stow_item?).and_return(true)
+
+      instance.send(:set_warhorn_or_egg)
+
+      expect(instance.instance_variable_get(:@warhorn_or_egg)).to eq(['egg'])
+    end
+
+    it 'warns when no items are found at all' do
+      instance = build_ability_process(egg_count: 1, warhorn_nouns: ['warhorn'])
+      allow(DRCI).to receive(:get_item?).and_return(false)
+      allow(DRCI).to receive(:remove_item?).and_return(false)
+
+      instance.send(:set_warhorn_or_egg)
+
+      expect(instance.instance_variable_get(:@warhorn_or_egg)).to be_empty
+      expect(DRC).to have_received(:message).with(/No eggs or warhorns found/)
+    end
+
+    it 'warns when fewer eggs found than configured' do
+      call_count = 0
+      instance = build_ability_process(egg_count: 2, warhorn_nouns: [])
+      allow(DRCI).to receive(:get_item?) do |_arg|
+        call_count += 1
+        if call_count == 1
+          stub_right_hand_with_id('10')
+          true
+        else
+          false
+        end
+      end
+      allow(DRCI).to receive(:stow_item?).and_return(true)
+
+      instance.send(:set_warhorn_or_egg)
+
+      expect(DRC).to have_received(:message).with(/wanted 2 egg.*only found 1/)
+    end
+  end
+
+  # ===========================================================================
+  # #use_warhorn_or_egg -- room effect gate
+  # ===========================================================================
+  describe '#use_warhorn_or_egg' do
+    it 'skips when room effect is still active (< 600s)' do
+      UserVars.warhorn = { "last_warhorn_or_egg" => Time.now - 300 }
+      instance = build_ability_process(warhorn_or_egg: ['egg'], egg_ids: ['10'])
+      game_state = build_game_state
+
+      instance.send(:use_warhorn_or_egg, game_state)
+
+      expect(DRC).not_to have_received(:bput).with(/invoke/, anything, anything, anything, anything, anything)
+    end
+
+    it 'attempts use when room effect has expired (>= 600s)' do
+      UserVars.warhorn = { "last_warhorn_or_egg" => Time.now - 601 }
+      instance = build_ability_process(
+        warhorn_or_egg: ['egg'],
+        egg_ids: ['10'],
+        item_cooldowns: {}
+      )
+      game_state = build_game_state
+      allow(DRC).to receive(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+                                  .and_return('light envelops the area briefly')
+
+      instance.send(:use_warhorn_or_egg, game_state)
+
+      expect(DRC).to have_received(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+    end
+
+    it 'rotates the type after each call' do
+      UserVars.warhorn = { "last_warhorn_or_egg" => Time.now - 601 }
+      instance = build_ability_process(
+        warhorn_or_egg: %w[egg warhorn],
+        egg_ids: ['10'],
+        warhorn_items: [{ id: '20', worn: false }],
+        item_cooldowns: {}
+      )
+      game_state = build_game_state
+      allow(DRC).to receive(:bput).and_return('light envelops the area briefly')
+
+      instance.send(:use_warhorn_or_egg, game_state)
+
+      expect(instance.instance_variable_get(:@warhorn_or_egg)).to eq(%w[warhorn egg])
+    end
+  end
+
+  # ===========================================================================
+  # #use_egg? -- per-item cooldown and error handling
+  # ===========================================================================
+  describe '#use_egg?' do
+    it 'returns true on successful invocation' do
+      instance = build_ability_process(egg_ids: ['10'], item_cooldowns: {})
+      allow(DRC).to receive(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+                                  .and_return('light envelops the area briefly')
+
+      expect(instance.send(:use_egg?)).to be true
+    end
+
+    it 'records cooldown timestamp on success' do
+      instance = build_ability_process(egg_ids: ['10'], item_cooldowns: {})
+      allow(DRC).to receive(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+                                  .and_return('light envelops the area briefly')
+
+      instance.send(:use_egg?)
+
+      cooldowns = instance.instance_variable_get(:@item_cooldowns)
+      expect(cooldowns['10']).to be_within(2).of(Time.now)
+    end
+
+    it 'skips egg on cooldown and tries the next one' do
+      instance = build_ability_process(
+        egg_ids: %w[10 20],
+        item_cooldowns: { '10' => Time.now }
+      )
+      allow(DRC).to receive(:bput).with("invoke #20", anything, anything, anything, anything, anything)
+                                  .and_return('light envelops the area briefly')
+
+      expect(instance.send(:use_egg?)).to be true
+      expect(DRC).not_to have_received(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+    end
+
+    it 'returns false and removes egg type when area inhibits' do
+      rotation = %w[egg warhorn]
+      instance = build_ability_process(
+        egg_ids: ['10'],
+        item_cooldowns: {},
+        warhorn_or_egg: rotation
+      )
+      allow(DRC).to receive(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+                                  .and_return('Something about the area inhibits')
+
+      result = instance.send(:use_egg?)
+
+      expect(result).to be false
+      expect(rotation).not_to include('egg')
+    end
+
+    it 'removes a missing egg from the list and tries remaining' do
+      instance = build_ability_process(
+        egg_ids: %w[10 20],
+        item_cooldowns: {},
+        warhorn_or_egg: ['egg']
+      )
+      allow(DRC).to receive(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+                                  .and_return('Invoke what?')
+      allow(DRC).to receive(:bput).with("invoke #20", anything, anything, anything, anything, anything)
+                                  .and_return('light envelops the area briefly')
+
+      expect(instance.send(:use_egg?)).to be true
+      expect(instance.instance_variable_get(:@egg_ids)).to eq(['20'])
+    end
+
+    it 'returns false when all eggs are missing' do
+      instance = build_ability_process(
+        egg_ids: ['10'],
+        item_cooldowns: {},
+        warhorn_or_egg: ['egg']
+      )
+      allow(DRC).to receive(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+                                  .and_return('Invoke what?')
+
+      expect(instance.send(:use_egg?)).to be false
+    end
+
+    it 'sets a 60s retry cooldown when egg is dim/sluggish' do
+      instance = build_ability_process(egg_ids: ['10'], item_cooldowns: {})
+      allow(DRC).to receive(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+                                  .and_return('The red light within the egg is dim and moves about sluggishly')
+
+      instance.send(:use_egg?)
+
+      cooldown = instance.instance_variable_get(:@item_cooldowns)['10']
+      expect(cooldown).to be_within(2).of(Time.now - 900 + 60)
+    end
+
+    it 'returns false when hidden and cannot use egg' do
+      instance = build_ability_process(egg_ids: ['10'], item_cooldowns: {})
+      allow(DRC).to receive(:bput).with("invoke #10", anything, anything, anything, anything, anything)
+                                  .and_return('You cannot stay hidden while using the egg.')
+
+      expect(instance.send(:use_egg?)).to be false
+    end
+
+    it 'returns false when egg_ids is empty' do
+      instance = build_ability_process(egg_ids: [], item_cooldowns: {})
+
+      expect(instance.send(:use_egg?)).to be false
+    end
+
+    it 'returns false when all eggs are on cooldown' do
+      instance = build_ability_process(
+        egg_ids: %w[10 20],
+        item_cooldowns: { '10' => Time.now, '20' => Time.now }
+      )
+
+      expect(instance.send(:use_egg?)).to be false
+    end
+  end
+
+  # ===========================================================================
+  # #use_warhorn? -- per-item cooldown and error handling
+  # ===========================================================================
+  describe '#use_warhorn?' do
+    let(:game_state) { build_game_state }
+
+    it 'returns true on successful exhale' do
+      instance = build_ability_process(
+        warhorn_items: [{ id: '20', worn: false }],
+        item_cooldowns: {}
+      )
+      allow(DRC).to receive(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+                                  .and_return('You get a silver warhorn.')
+      allow(DRC).to receive(:bput).with("exhale #20 lure", anything, anything, anything, anything)
+                                  .and_return('You sound a series of bursts from the')
+      allow(instance).to receive(:waitrt?)
+      allow(DRC).to receive(:bput).with("stow #20", anything, anything, anything, anything)
+                                  .and_return('You put')
+
+      expect(instance.send(:use_warhorn?, game_state)).to be true
+    end
+
+    it 'uses remove verb for worn warhorns' do
+      instance = build_ability_process(
+        warhorn_items: [{ id: '20', worn: true }],
+        item_cooldowns: {}
+      )
+      allow(DRC).to receive(:bput).with("remove #20", anything, anything, anything, anything, anything, anything)
+                                  .and_return('You remove a silver warhorn.')
+      allow(DRC).to receive(:bput).with("exhale #20 lure", anything, anything, anything, anything)
+                                  .and_return('You sound a series of bursts from the')
+      allow(instance).to receive(:waitrt?)
+      allow(DRC).to receive(:bput).with("wear #20", anything, anything, anything, anything)
+                                  .and_return('You attach')
+
+      expect(instance.send(:use_warhorn?, game_state)).to be true
+      expect(DRC).to have_received(:bput).with("remove #20", anything, anything, anything, anything, anything, anything)
+    end
+
+    it 'skips warhorn on cooldown and tries the next one' do
+      instance = build_ability_process(
+        warhorn_items: [{ id: '20', worn: false }, { id: '30', worn: false }],
+        item_cooldowns: { '20' => Time.now }
+      )
+      allow(DRC).to receive(:bput).with("get #30", anything, anything, anything, anything, anything, anything)
+                                  .and_return('You get')
+      allow(DRC).to receive(:bput).with("exhale #30 lure", anything, anything, anything, anything)
+                                  .and_return('You sound a series of bursts from the')
+      allow(instance).to receive(:waitrt?)
+      allow(DRC).to receive(:bput).with("stow #30", anything, anything, anything, anything)
+                                  .and_return('You put')
+
+      expect(instance.send(:use_warhorn?, game_state)).to be true
+      expect(DRC).not_to have_received(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+    end
+
+    it 'sets a 60s retry cooldown when lungs are tired' do
+      instance = build_ability_process(
+        warhorn_items: [{ id: '20', worn: false }],
+        item_cooldowns: {},
+        warhorn_cooldown: 1200
+      )
+      allow(DRC).to receive(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+                                  .and_return('You get')
+      allow(DRC).to receive(:bput).with("exhale #20 lure", anything, anything, anything, anything)
+                                  .and_return('Your lungs are tired from having sounded a')
+      allow(DRC).to receive(:bput).with("stow #20", anything, anything, anything, anything)
+                                  .and_return('You put')
+
+      instance.send(:use_warhorn?, game_state)
+
+      cooldown = instance.instance_variable_get(:@item_cooldowns)['20']
+      expect(cooldown).to be_within(2).of(Time.now - 1200 + 60)
+    end
+
+    it 'returns false and removes warhorn type when area inhibits' do
+      rotation = %w[warhorn egg]
+      instance = build_ability_process(
+        warhorn_items: [{ id: '20', worn: false }],
+        item_cooldowns: {},
+        warhorn_or_egg: rotation
+      )
+      allow(DRC).to receive(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+                                  .and_return('You get')
+      allow(DRC).to receive(:bput).with("exhale #20 lure", anything, anything, anything, anything)
+                                  .and_return('Something about the area inhibits')
+      allow(DRC).to receive(:bput).with("stow #20", anything, anything, anything, anything)
+                                  .and_return('You put')
+
+      result = instance.send(:use_warhorn?, game_state)
+
+      expect(result).to be false
+      expect(rotation).not_to include('warhorn')
+    end
+
+    it 'removes a missing warhorn from the list and tries remaining' do
+      item1 = { id: '20', worn: false }
+      item2 = { id: '30', worn: false }
+      instance = build_ability_process(
+        warhorn_items: [item1, item2],
+        item_cooldowns: {}
+      )
+      allow(DRC).to receive(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+                                  .and_return('What were you referring to')
+      allow(DRC).to receive(:bput).with("get #30", anything, anything, anything, anything, anything, anything)
+                                  .and_return('You get')
+      allow(DRC).to receive(:bput).with("exhale #30 lure", anything, anything, anything, anything)
+                                  .and_return('You sound a series of bursts from the')
+      allow(instance).to receive(:waitrt?)
+      allow(DRC).to receive(:bput).with("stow #30", anything, anything, anything, anything)
+                                  .and_return('You put')
+
+      expect(instance.send(:use_warhorn?, game_state)).to be true
+      expect(instance.instance_variable_get(:@warhorn_items)).not_to include(item1)
+    end
+
+    it 'returns false when hands are full' do
+      instance = build_ability_process(
+        warhorn_items: [{ id: '20', worn: false }],
+        item_cooldowns: {}
+      )
+      allow(DRC).to receive(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+                                  .and_return('You need a free hand')
+
+      expect(instance.send(:use_warhorn?, game_state)).to be false
+    end
+
+    it 'returns false when all warhorns are on cooldown' do
+      instance = build_ability_process(
+        warhorn_items: [{ id: '20', worn: false }, { id: '30', worn: false }],
+        item_cooldowns: { '20' => Time.now, '30' => Time.now }
+      )
+
+      expect(instance.send(:use_warhorn?, game_state)).to be false
+    end
+
+    it 'returns false and removes warhorn type when player cannot use warhorns' do
+      rotation = %w[warhorn egg]
+      instance = build_ability_process(
+        warhorn_items: [{ id: '20', worn: false }],
+        item_cooldowns: {},
+        warhorn_or_egg: rotation
+      )
+      allow(DRC).to receive(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+                                  .and_return('You get')
+      allow(DRC).to receive(:bput).with("exhale #20 lure", anything, anything, anything, anything)
+                                  .and_return('not accomplishing much and looking rather silly')
+      allow(DRC).to receive(:bput).with("stow #20", anything, anything, anything, anything)
+                                  .and_return('You put')
+
+      result = instance.send(:use_warhorn?, game_state)
+
+      expect(result).to be false
+      expect(rotation).not_to include('warhorn')
+    end
+
+    it 'wields whirlwind offhand when all warhorns exhausted' do
+      instance = build_ability_process(
+        warhorn_items: [],
+        item_cooldowns: {}
+      )
+
+      instance.send(:use_warhorn?, game_state)
+
+      expect(game_state).to have_received(:wield_whirlwind_offhand)
+    end
+  end
+
+  # ===========================================================================
+  # #stow_warhorn_item
+  # ===========================================================================
+  describe '#stow_warhorn_item' do
+    it 'uses stow for non-worn items' do
+      instance = build_ability_process
+      allow(DRC).to receive(:bput).and_return('You put')
+
+      instance.send(:stow_warhorn_item, { id: '20', worn: false })
+
+      expect(DRC).to have_received(:bput).with('stow #20', anything, anything, anything, anything)
+    end
+
+    it 'uses wear for worn items' do
+      instance = build_ability_process
+      allow(DRC).to receive(:bput).and_return('You attach')
+
+      instance.send(:stow_warhorn_item, { id: '20', worn: true })
+
+      expect(DRC).to have_received(:bput).with('wear #20', anything, anything, anything, anything)
+    end
+  end
+
+  # ===========================================================================
+  # Bad YAML config parsing -- tests the case expressions in initialize
+  # that produce @warhorn_nouns and @egg_count from raw settings values.
+  #
+  # We can't call initialize (needs full game I/O), so we replicate the
+  # case expressions inline and verify the derived values fed to downstream
+  # methods behave correctly.
+  # ===========================================================================
+  describe 'bad YAML config edge cases' do
+    # Replicate the warhorn case expression from initialize
+    def warhorn_nouns_from(raw)
+      case raw
+      when Array then raw
+      when String then [raw]
+      when true then ['warhorn']
+      else []
+      end
+    end
+
+    # Replicate the egg case expression from initialize
+    def egg_count_from(raw)
+      case raw
+      when Integer then raw
+      when true, String then 1
+      else 0
+      end
+    end
+
+    # Replicate the guard that decides whether to call set_warhorn_or_egg
+    def should_setup?(nouns, count)
+      !(nouns.empty? && count < 1)
+    end
+
+    # =========================================================================
+    # warhorn config parsing
+    # =========================================================================
+    describe 'warhorn config parsing' do
+      it 'treats integer 42 as no warhorns' do
+        nouns = warhorn_nouns_from(42)
+        expect(nouns).to eq([])
+      end
+
+      it 'treats false as no warhorns' do
+        nouns = warhorn_nouns_from(false)
+        expect(nouns).to eq([])
+      end
+
+      it 'treats nil as no warhorns' do
+        nouns = warhorn_nouns_from(nil)
+        expect(nouns).to eq([])
+      end
+
+      it 'wraps a single string in an array' do
+        nouns = warhorn_nouns_from('silver warhorn')
+        expect(nouns).to eq(['silver warhorn'])
+      end
+
+      it 'passes an array through unchanged' do
+        nouns = warhorn_nouns_from(%w[warhorn horn])
+        expect(nouns).to eq(%w[warhorn horn])
+      end
+
+      it 'treats true as default warhorn noun' do
+        nouns = warhorn_nouns_from(true)
+        expect(nouns).to eq(['warhorn'])
+      end
+
+      it 'passes an empty array through (no warhorns)' do
+        nouns = warhorn_nouns_from([])
+        expect(nouns).to eq([])
+      end
+
+      it 'passes an array with non-string elements through without filtering' do
+        nouns = warhorn_nouns_from([true, 42, 'warhorn'])
+        expect(nouns).to eq([true, 42, 'warhorn'])
+      end
+    end
+
+    # =========================================================================
+    # egg config parsing
+    # =========================================================================
+    describe 'egg config parsing' do
+      it 'treats integer 0 as zero eggs' do
+        expect(egg_count_from(0)).to eq(0)
+      end
+
+      it 'treats negative integer as negative count' do
+        expect(egg_count_from(-1)).to eq(-1)
+      end
+
+      it 'treats integer 2 as two eggs' do
+        expect(egg_count_from(2)).to eq(2)
+      end
+
+      it 'treats integer 3 as three (even though only 2 ordinals supported)' do
+        expect(egg_count_from(3)).to eq(3)
+      end
+
+      it 'treats true as 1 egg' do
+        expect(egg_count_from(true)).to eq(1)
+      end
+
+      it 'treats a string as 1 egg' do
+        expect(egg_count_from('yes')).to eq(1)
+      end
+
+      it 'treats false as 0 eggs' do
+        expect(egg_count_from(false)).to eq(0)
+      end
+
+      it 'treats nil as 0 eggs' do
+        expect(egg_count_from(nil)).to eq(0)
+      end
+
+      it 'treats an array as 0 eggs' do
+        expect(egg_count_from([1, 2])).to eq(0)
+      end
+
+      it 'treats a hash as 0 eggs' do
+        expect(egg_count_from({ count: 2 })).to eq(0)
+      end
+
+      it 'treats float 1.5 as 0 eggs (not Integer)' do
+        expect(egg_count_from(1.5)).to eq(0)
+      end
+    end
+
+    # =========================================================================
+    # setup guard -- should set_warhorn_or_egg be called?
+    # =========================================================================
+    describe 'setup guard' do
+      it 'skips setup when both empty/zero' do
+        expect(should_setup?([], 0)).to be false
+      end
+
+      it 'runs setup when warhorn nouns present but egg_count is 0' do
+        expect(should_setup?(['warhorn'], 0)).to be true
+      end
+
+      it 'runs setup when egg_count is 1 but no warhorn nouns' do
+        expect(should_setup?([], 1)).to be true
+      end
+
+      it 'runs setup when both present' do
+        expect(should_setup?(['warhorn'], 2)).to be true
+      end
+
+      it 'runs setup when egg_count is negative (will discover 0 eggs)' do
+        expect(should_setup?([], -1)).to be false
+      end
+    end
+
+    # =========================================================================
+    # set_warhorn_or_egg with bad config-derived values
+    # =========================================================================
+    describe 'set_warhorn_or_egg with degenerate configs' do
+      it 'handles egg_count 0 with warhorn_nouns present (warhorn only)' do
+        instance = build_ability_process(egg_count: 0, warhorn_nouns: ['warhorn'])
+        stub_right_hand_with_id('10')
+        allow(DRCI).to receive(:remove_item?).and_return(true)
+        allow(DRCI).to receive(:wear_item?).and_return(true)
+
+        instance.send(:set_warhorn_or_egg)
+
+        expect(instance.instance_variable_get(:@warhorn_or_egg)).to eq(['warhorn'])
+        expect(instance.instance_variable_get(:@egg_ids)).to be_empty
+      end
+
+      it 'handles warhorn_nouns with non-string elements gracefully' do
+        instance = build_ability_process(egg_count: 0, warhorn_nouns: [true, 42])
+        allow(DRCI).to receive(:remove_item?).and_return(false)
+        allow(DRCI).to receive(:get_item?).and_return(false)
+
+        instance.send(:set_warhorn_or_egg)
+
+        expect(instance.instance_variable_get(:@warhorn_or_egg)).to be_empty
+        expect(DRC).to have_received(:message).with(/No eggs or warhorns found/)
+      end
+
+      it 'handles egg_count 3 (only discovers first 2, skips unsupported ordinal)' do
+        instance = build_ability_process(egg_count: 3, warhorn_nouns: [])
+        call_count = 0
+        allow(DRCI).to receive(:get_item?) do |_arg|
+          call_count += 1
+          stub_right_hand_with_id("e#{call_count}")
+          true
+        end
+        allow(DRCI).to receive(:stow_item?).and_return(true)
+
+        instance.send(:set_warhorn_or_egg)
+
+        # Only 2 ordinals are supported ("egg" and "second egg")
+        expect(instance.instance_variable_get(:@egg_ids).size).to eq(2)
+        expect(DRC).to have_received(:message).with(/wanted 3 egg.*only found 2/)
+      end
+
+      it 'handles empty warhorn_nouns array (no discovery attempted)' do
+        instance = build_ability_process(egg_count: 1, warhorn_nouns: [])
+        stub_right_hand_with_id('10')
+        allow(DRCI).to receive(:get_item?).and_return(true)
+        allow(DRCI).to receive(:stow_item?).and_return(true)
+
+        instance.send(:set_warhorn_or_egg)
+
+        expect(instance.instance_variable_get(:@warhorn_items)).to be_empty
+        expect(instance.instance_variable_get(:@warhorn_or_egg)).to eq(['egg'])
+      end
+    end
+
+    # =========================================================================
+    # use methods with zero/negative warhorn_cooldown
+    # =========================================================================
+    describe 'warhorn_cooldown edge cases' do
+      let(:game_state) { build_game_state }
+
+      it 'warhorn_cooldown 0 means cooldown expires immediately' do
+        instance = build_ability_process(
+          warhorn_items: [{ id: '20', worn: false }],
+          item_cooldowns: { '20' => Time.now - 1 },
+          warhorn_cooldown: 0
+        )
+        allow(DRC).to receive(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+                                    .and_return('You get')
+        allow(DRC).to receive(:bput).with("exhale #20 lure", anything, anything, anything, anything)
+                                    .and_return('You sound a series of bursts from the')
+        allow(instance).to receive(:waitrt?)
+        allow(DRC).to receive(:bput).with("stow #20", anything, anything, anything, anything)
+                                    .and_return('You put')
+
+        expect(instance.send(:use_warhorn?, game_state)).to be true
+      end
+
+      it 'negative warhorn_cooldown means cooldown is always expired' do
+        instance = build_ability_process(
+          warhorn_items: [{ id: '20', worn: false }],
+          item_cooldowns: { '20' => Time.now },
+          warhorn_cooldown: -500
+        )
+        allow(DRC).to receive(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+                                    .and_return('You get')
+        allow(DRC).to receive(:bput).with("exhale #20 lure", anything, anything, anything, anything)
+                                    .and_return('You sound a series of bursts from the')
+        allow(instance).to receive(:waitrt?)
+        allow(DRC).to receive(:bput).with("stow #20", anything, anything, anything, anything)
+                                    .and_return('You put')
+
+        expect(instance.send(:use_warhorn?, game_state)).to be true
+      end
+
+      it 'lungs-tired retry with cooldown 0 sets retry ~60s from now' do
+        instance = build_ability_process(
+          warhorn_items: [{ id: '20', worn: false }],
+          item_cooldowns: {},
+          warhorn_cooldown: 0
+        )
+        allow(DRC).to receive(:bput).with("get #20", anything, anything, anything, anything, anything, anything)
+                                    .and_return('You get')
+        allow(DRC).to receive(:bput).with("exhale #20 lure", anything, anything, anything, anything)
+                                    .and_return('Your lungs are tired from having sounded a')
+        allow(DRC).to receive(:bput).with("stow #20", anything, anything, anything, anything)
+                                    .and_return('You put')
+
+        instance.send(:use_warhorn?, game_state)
+
+        cooldown = instance.instance_variable_get(:@item_cooldowns)['20']
+        # Time.now - 0 + 60 = ~60s from now
+        expect(cooldown).to be_within(2).of(Time.now + 60)
+      end
+    end
+
+    # =========================================================================
+    # Concurrent removal of all items during use
+    # =========================================================================
+    describe 'all items vanish during use' do
+      let(:game_state) { build_game_state }
+
+      it 'handles all eggs disappearing one by one' do
+        instance = build_ability_process(
+          egg_ids: %w[10 20 30],
+          item_cooldowns: {},
+          warhorn_or_egg: ['egg']
+        )
+        allow(DRC).to receive(:bput).with(/invoke #/, anything, anything, anything, anything, anything)
+                                    .and_return('Invoke what?')
+
+        expect(instance.send(:use_egg?)).to be false
+        expect(instance.instance_variable_get(:@egg_ids)).to be_empty
+      end
+
+      it 'handles all warhorns disappearing one by one' do
+        instance = build_ability_process(
+          warhorn_items: [
+            { id: '20', worn: false },
+            { id: '30', worn: false },
+            { id: '40', worn: false }
+          ],
+          item_cooldowns: {}
+        )
+        allow(DRC).to receive(:bput).with(/get #/, anything, anything, anything, anything, anything, anything)
+                                    .and_return('What were you referring to')
+
+        expect(instance.send(:use_warhorn?, game_state)).to be false
+        expect(instance.instance_variable_get(:@warhorn_items)).to be_empty
+      end
+    end
+
+    # =========================================================================
+    # Mixed success/failure across multiple items
+    # =========================================================================
+    describe 'mixed item states' do
+      it 'first egg cooldown, second egg missing, third egg succeeds' do
+        instance = build_ability_process(
+          egg_ids: %w[10 20 30],
+          item_cooldowns: { '10' => Time.now },
+          warhorn_or_egg: ['egg']
+        )
+        allow(DRC).to receive(:bput).with("invoke #20", anything, anything, anything, anything, anything)
+                                    .and_return('Invoke what?')
+        allow(DRC).to receive(:bput).with("invoke #30", anything, anything, anything, anything, anything)
+                                    .and_return('light envelops the area briefly')
+
+        expect(instance.send(:use_egg?)).to be true
+        expect(instance.instance_variable_get(:@egg_ids)).to eq(%w[10 30])
+      end
+
+      it 'first warhorn cooldown, second warhorn lungs-tired, all exhausted' do
+        instance = build_ability_process(
+          warhorn_items: [
+            { id: '20', worn: false },
+            { id: '30', worn: false }
+          ],
+          item_cooldowns: { '20' => Time.now },
+          warhorn_cooldown: 1200
+        )
+        allow(DRC).to receive(:bput).with("get #30", anything, anything, anything, anything, anything, anything)
+                                    .and_return('You get')
+        allow(DRC).to receive(:bput).with("exhale #30 lure", anything, anything, anything, anything)
+                                    .and_return('Your lungs are tired from having sounded a')
+        allow(DRC).to receive(:bput).with("stow #30", anything, anything, anything, anything)
+                                    .and_return('You put')
+
+        game_state = build_game_state
+        expect(instance.send(:use_warhorn?, game_state)).to be false
+        expect(instance.instance_variable_get(:@item_cooldowns)['30']).not_to be_nil
+      end
+    end
+  end
+end
+
+# ###################################################################
+# MERGED FROM spec/combat_trainer_slivers_spec.rb
+# ###################################################################
+
+# ===========================================================================
+# SpellProcess#check_slivers -- sliver detection and creation for Moon Mages
+# ===========================================================================
+RSpec.describe SpellProcess do
+  # Build a SpellProcess without calling initialize
+  def build_spell_process(**overrides)
+    instance = SpellProcess.allocate
+    defaults = {
+      tk_spell: { 'abbrev' => 'tkt' },
+      tk_ammo: nil,
+      settings: OpenStruct.new
+    }
+    defaults.merge(overrides).each do |k, v|
+      instance.instance_variable_set(:"@#{k}", v)
+    end
+    instance
+  end
+
+  def build_game_state(**attrs)
+    defaults = { casting: false }
+    double('GameState', defaults.merge(attrs))
+  end
+
+  def setup_moon_mage_with_moonblade
+    DRStats.guild = 'Moon Mage'
+    DRSpells._set_known_spells({ 'Moonblade' => true })
+    UserVars._set_moons({ 'visible' => ['Katamba'] })
+    # Stub get_data to return spell data with Moonblade
+    $test_data = OpenStruct.new(
+      spells: OpenStruct.new(
+        spell_data: { 'Moonblade' => { 'mana' => 5, 'prep_time' => 5 } }
+      )
+    )
+  end
+
+  describe '#check_slivers' do
+    context 'guard clauses' do
+      it 'returns early if character does not know Moonblade' do
+        DRStats.guild = 'Moon Mage'
+        DRSpells._set_known_spells({})
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        expect(DRCA).not_to receive(:cast_spell)
+        instance.send(:check_slivers, game_state)
+      end
+
+      it 'returns early if character is not a Moon Mage' do
+        DRStats.guild = 'Warrior Mage'
+        DRSpells._set_known_spells({ 'Moonblade' => true })
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        expect(DRCA).not_to receive(:cast_spell)
+        instance.send(:check_slivers, game_state)
+      end
+
+      it 'returns early if no TK spell is configured' do
+        DRStats.guild = 'Moon Mage'
+        DRSpells._set_known_spells({ 'Moonblade' => true })
+
+        instance = build_spell_process(tk_spell: nil)
+        game_state = build_game_state
+
+        expect(DRCA).not_to receive(:cast_spell)
+        instance.send(:check_slivers, game_state)
+      end
+
+      it 'returns early if already casting' do
+        DRStats.guild = 'Moon Mage'
+        DRSpells._set_known_spells({ 'Moonblade' => true })
+
+        instance = build_spell_process
+        game_state = build_game_state(casting: true)
+
+        expect(DRCA).not_to receive(:cast_spell)
+        instance.send(:check_slivers, game_state)
+      end
+
+      it 'returns early if slivers already exist' do
+        DRStats.guild = 'Moon Mage'
+        DRSpells._set_known_spells({ 'Moonblade' => true })
+        DRSpells._set_slivers(true)
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        expect(DRCA).not_to receive(:cast_spell)
+        instance.send(:check_slivers, game_state)
+      end
+
+      it 'returns early if no moons are visible' do
+        DRStats.guild = 'Moon Mage'
+        DRSpells._set_known_spells({ 'Moonblade' => true })
+        UserVars._set_moons({ 'visible' => [] })
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        expect(DRCA).not_to receive(:cast_spell)
+        instance.send(:check_slivers, game_state)
+      end
+    end
+
+    context 'when slivers need to be created' do
+      before(:each) do
+        setup_moon_mage_with_moonblade
+        allow(DRCA).to receive(:cast_spell)
+      end
+
+      it 'casts moonblade and breaks it on success' do
+        allow(DRC).to receive(:bput)
+          .with('break moonblade', 'The slivers drift about', 'dissipate without any benefit', 'Break what?')
+          .and_return('The slivers drift about')
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        instance.send(:check_slivers, game_state)
+
+        expect(DRCA).to have_received(:cast_spell).once
+        expect(DRC).to have_received(:bput).with('break moonblade', anything, anything, anything).once
+      end
+
+      it 'retries up to 3 times on failure' do
+        allow(DRC).to receive(:bput)
+          .with('break moonblade', 'The slivers drift about', 'dissipate without any benefit', 'Break what?')
+          .and_return('dissipate without any benefit', 'dissipate without any benefit', 'dissipate without any benefit')
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        instance.send(:check_slivers, game_state)
+
+        expect(DRCA).to have_received(:cast_spell).exactly(3).times
+        expect(DRC).to have_received(:bput).with('break moonblade', anything, anything, anything).exactly(3).times
+      end
+
+      it 'stops retrying after first success' do
+        allow(DRC).to receive(:bput)
+          .with('break moonblade', 'The slivers drift about', 'dissipate without any benefit', 'Break what?')
+          .and_return('dissipate without any benefit', 'The slivers drift about')
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        instance.send(:check_slivers, game_state)
+
+        expect(DRCA).to have_received(:cast_spell).exactly(2).times
+      end
+
+      it 'logs failure message when all retries are exhausted' do
+        allow(DRC).to receive(:bput)
+          .with('break moonblade', 'The slivers drift about', 'dissipate without any benefit', 'Break what?')
+          .and_return('dissipate without any benefit')
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        expect(DRC).to receive(:message).with(/Failed to create slivers.*3 attempts/)
+        instance.send(:check_slivers, game_state)
+      end
+
+      it 'does not log failure message on success' do
+        allow(DRC).to receive(:bput)
+          .with('break moonblade', 'The slivers drift about', 'dissipate without any benefit', 'Break what?')
+          .and_return('The slivers drift about')
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        expect(DRC).not_to receive(:message).with(/Failed to create slivers/)
+        instance.send(:check_slivers, game_state)
+      end
+    end
+
+    context 'prep time based on Lunar Magic rank' do
+      before(:each) do
+        setup_moon_mage_with_moonblade
+        allow(DRC).to receive(:bput)
+          .with('break moonblade', anything, anything, anything)
+          .and_return('The slivers drift about')
+      end
+
+      it 'uses prep_time 1 for Lunar Magic >= 400' do
+        DRSkill._set_rank('Lunar Magic', 450)
+        allow(DRCA).to receive(:cast_spell)
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        instance.send(:check_slivers, game_state)
+
+        expect(DRCA).to have_received(:cast_spell) do |spell_data, _settings|
+          expect(spell_data['prep_time']).to eq(1)
+        end
+      end
+
+      it 'uses prep_time 2 for Lunar Magic 300-399' do
+        DRSkill._set_rank('Lunar Magic', 350)
+        allow(DRCA).to receive(:cast_spell)
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        instance.send(:check_slivers, game_state)
+
+        expect(DRCA).to have_received(:cast_spell) do |spell_data, _settings|
+          expect(spell_data['prep_time']).to eq(2)
+        end
+      end
+
+      it 'uses prep_time 3 for Lunar Magic 200-299' do
+        DRSkill._set_rank('Lunar Magic', 250)
+        allow(DRCA).to receive(:cast_spell)
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        instance.send(:check_slivers, game_state)
+
+        expect(DRCA).to have_received(:cast_spell) do |spell_data, _settings|
+          expect(spell_data['prep_time']).to eq(3)
+        end
+      end
+
+      it 'does not override prep_time for Lunar Magic < 200' do
+        DRSkill._set_rank('Lunar Magic', 150)
+        allow(DRCA).to receive(:cast_spell)
+
+        instance = build_spell_process
+        game_state = build_game_state
+
+        instance.send(:check_slivers, game_state)
+
+        expect(DRCA).to have_received(:cast_spell) do |spell_data, _settings|
+          # prep_time should remain at the spell data default (5)
+          expect(spell_data['prep_time']).to eq(5)
+        end
+      end
+    end
+  end
+end
+
+# ###################################################################
+# MERGED FROM spec/combat_trainer_gempouch_spec.rb
+# ###################################################################
+
+# ===========================================================================
+# LootProcess#stow_loot -- gem pouch swap when pouch is full
+#
+# Flow: stow_loot tries to stow an item. If the pouch-full flag fires
+# (set by a game message matcher), the method drops the item, swaps
+# the full pouch for a spare via DRCI.swap_out_full_gempouch?, then
+# picks up the dropped gem.
+# ===========================================================================
+RSpec.describe LootProcess do
+  def build_loot_process(**overrides)
+    instance = LootProcess.allocate
+    defaults = {
+      tie_bundle: false,
+      skin: false,
+      dissect: false,
+      dump_timer: Time.now,
+      dump_junk: false,
+      dump_item_count: 10,
+      autoloot_container: nil,
+      autoloot_gems: false,
+      loot_bodies: true,
+      lootables: [],
+      gem_nouns: ['diamond'],
+      box_nouns: [],
+      box_loot_limit: nil,
+      current_box_count: 0,
+      loot_specials: [],
+      gem_pouch_adjective: 'black',
+      gem_pouch_noun: 'pouch',
+      full_pouch_container: 'backpack',
+      spare_gem_pouch_container: 'locker',
+      tie_pouch: false,
+      equipment_manager: double('EquipmentManager', stow_weapon: nil, wield_weapon?: nil, is_listed_item?: false)
+    }
+    defaults.merge(overrides).each do |k, v|
+      instance.instance_variable_set(:"@#{k}", v)
+    end
+    instance
+  end
+
+  def build_game_state(**attrs)
+    defaults = {
+      need_bundle: false,
+      mob_died: false,
+      npcs: [],
+      skinnable?: false,
+      finish_killing?: false,
+      finish_spell_casting?: false,
+      stowing?: false,
+      currently_whirlwinding: false
+    }
+    state = double('GameState', defaults.merge(attrs))
+    allow(state).to receive(:unlootable)
+    allow(state).to receive(:lootable?).and_return(true)
+    state
+  end
+
+  describe '#stow_loot (pouch-full swap)' do
+    before(:each) do
+      # Allow all bput calls by default (stow, drop, etc.)
+      allow(DRC).to receive(:bput).and_return('You put')
+      allow(DRCI).to receive(:swap_out_full_gempouch?).and_return(true)
+      allow(DRCI).to receive(:get_item_unsafe).and_return(false)
+    end
+
+    context 'when pouch-full flag is set and swap succeeds' do
+      let(:game_state) { build_game_state }
+
+      before(:each) do
+        Flags['container-full'] = nil
+        # The pouch-full flag fires as a side effect during the stow bput call.
+        # Simulate this by having the stow bput set the flag.
+        allow(DRC).to receive(:bput).with(/^stow /, *Array.new(15, anything)) do
+          Flags['pouch-full'] = true
+          'You put'
+        end
+      end
+
+      it 'calls DRCI.swap_out_full_gempouch? with correct arguments' do
+        instance = build_loot_process
+        instance.send(:stow_loot, 'diamond', game_state)
+
+        expect(DRCI).to have_received(:swap_out_full_gempouch?).with(
+          'black', 'pouch', 'backpack', 'locker', false
+        )
+      end
+
+      it 'passes tie_pouch=true when configured' do
+        instance = build_loot_process(tie_pouch: true)
+        instance.send(:stow_loot, 'diamond', game_state)
+
+        expect(DRCI).to have_received(:swap_out_full_gempouch?).with(
+          'black', 'pouch', 'backpack', 'locker', true
+        )
+      end
+
+      it 'picks up the dropped gem after successful swap' do
+        instance = build_loot_process
+        instance.send(:stow_loot, 'diamond', game_state)
+
+        expect(DRC).to have_received(:bput).with('stow gem', anything, anything, anything, anything, anything, anything, anything)
+      end
+
+      it 'does not mark item as unlootable' do
+        instance = build_loot_process
+        instance.send(:stow_loot, 'diamond', game_state)
+
+        expect(game_state).not_to have_received(:unlootable)
+      end
+    end
+
+    context 'when pouch-full flag is set and swap fails' do
+      let(:game_state) { build_game_state }
+
+      before(:each) do
+        Flags['container-full'] = nil
+        allow(DRC).to receive(:bput).with(/^stow /, *Array.new(15, anything)) do
+          Flags['pouch-full'] = true
+          'You put'
+        end
+        allow(DRCI).to receive(:swap_out_full_gempouch?).and_return(false)
+      end
+
+      it 'marks item as unlootable' do
+        instance = build_loot_process
+        instance.send(:stow_loot, 'diamond', game_state)
+
+        expect(game_state).to have_received(:unlootable).with('diamond')
+      end
+
+      it 'does not try to pick up the gem' do
+        instance = build_loot_process
+        instance.send(:stow_loot, 'diamond', game_state)
+
+        expect(DRC).not_to have_received(:bput).with('stow gem', anything, anything, anything, anything, anything, anything, anything)
+      end
+    end
+
+    context 'when pouch-full flag is set but no spare container configured' do
+      let(:game_state) { build_game_state }
+
+      before(:each) do
+        Flags['container-full'] = nil
+        allow(DRC).to receive(:bput).with(/^stow /, *Array.new(15, anything)) do
+          Flags['pouch-full'] = true
+          'You put'
+        end
+      end
+
+      it 'marks item unlootable without attempting swap' do
+        instance = build_loot_process(spare_gem_pouch_container: nil)
+        instance.send(:stow_loot, 'diamond', game_state)
+
+        expect(game_state).to have_received(:unlootable).with('diamond')
+        expect(DRCI).not_to have_received(:swap_out_full_gempouch?)
+      end
+    end
+
+    context 'when pouch-full flag is not set' do
+      let(:game_state) { build_game_state }
+
+      before(:each) do
+        Flags['pouch-full'] = nil
+        Flags['container-full'] = nil
+      end
+
+      it 'does not attempt to swap pouches' do
+        instance = build_loot_process
+        instance.send(:stow_loot, 'diamond', game_state)
+
+        expect(DRCI).not_to have_received(:swap_out_full_gempouch?)
+      end
+    end
+  end
+end
+
+# ###################################################################
+# MERGED FROM spec/combat_trainer_force_cleanup_spec.rb
+# ###################################################################
+
+# ===========================================================================
+# GameState#force_cleanup specs
+#
+# Validates that force_cleanup advances the cleanup state machine past the
+# 'kill' phase, and is a no-op in all other states. This is the safety net
+# for when finishing the last mob takes too long (e.g. ranged weapons with
+# long aim cycles in multi-mob areas).
+# ===========================================================================
+RSpec.describe GameState do
+  def build_game_state(**overrides)
+    instance = GameState.allocate
+    defaults = {
+      clean_up_step: nil,
+      skip_last_kill: false,
+      stop_on_bleeding: false
+    }
+    defaults.merge(overrides).each do |k, v|
+      instance.instance_variable_set(:"@#{k}", v)
+    end
+    instance
+  end
+
+  describe '#force_cleanup' do
+    context 'when in the kill phase' do
+      it 'advances to clear_magic' do
+        gs = build_game_state(clean_up_step: 'kill')
+
+        gs.force_cleanup
+
+        expect(gs.finish_killing?).to be false
+        expect(gs.finish_spell_casting?).to be true
+      end
+
+      it 'is idempotent -- calling twice stays at clear_magic' do
+        gs = build_game_state(clean_up_step: 'kill')
+
+        gs.force_cleanup
+        gs.force_cleanup
+
+        expect(gs.finish_spell_casting?).to be true
+      end
+    end
+
+    # ------------------------------------------------------------------
+    # Adversarial: force_cleanup must not disrupt cleanup states that have
+    # already progressed past 'kill'. A bug here could skip stowing or
+    # cause the state machine to regress.
+    # ------------------------------------------------------------------
+    shared_examples 'no-op for non-kill phase' do |phase, description|
+      context "when in the #{description} phase (#{phase.inspect})" do
+        it 'does not change the cleanup step' do
+          gs = build_game_state(clean_up_step: phase)
+
+          gs.force_cleanup
+
+          expect(gs.instance_variable_get(:@clean_up_step)).to eq(phase)
+        end
+      end
+    end
+
+    include_examples 'no-op for non-kill phase', nil, 'not yet cleaning up'
+    include_examples 'no-op for non-kill phase', 'clear_magic', 'clear_magic'
+    include_examples 'no-op for non-kill phase', 'dismiss_pet', 'dismiss_pet'
+    include_examples 'no-op for non-kill phase', 'stow', 'stow'
+    include_examples 'no-op for non-kill phase', 'done', 'done'
+
+    # ------------------------------------------------------------------
+    # Adversarial: garbage or unexpected values must not be treated as
+    # 'kill'. The guard is an equality check, not a pattern match.
+    # ------------------------------------------------------------------
+    context 'when clean_up_step has an unexpected value' do
+      it 'does not change the cleanup step' do
+        gs = build_game_state(clean_up_step: 'bogus')
+
+        gs.force_cleanup
+
+        expect(gs.instance_variable_get(:@clean_up_step)).to eq('bogus')
+      end
+    end
+  end
+
+  # ===========================================================================
+  # next_clean_up_step interaction with force_cleanup
+  #
+  # Validates that the normal state machine and force_cleanup compose
+  # correctly -- force_cleanup mid-kill should allow normal progression
+  # to resume from clear_magic onward.
+  # ===========================================================================
+  describe '#next_clean_up_step after force_cleanup' do
+    it 'resumes normal progression from clear_magic through done' do
+      gs = build_game_state(clean_up_step: 'kill')
+
+      gs.force_cleanup
+      expect(gs.finish_spell_casting?).to be true
+
+      gs.next_clean_up_step
+      expect(gs.dismiss_pet?).to be true
+
+      gs.next_clean_up_step
+      expect(gs.stowing?).to be true
+
+      gs.next_clean_up_step
+      expect(gs.done_cleaning_up?).to be true
+    end
+  end
+
+  # ===========================================================================
+  # next_clean_up_step with skip_last_kill
+  #
+  # When skip_last_kill is true, next_clean_up_step skips 'kill' entirely.
+  # force_cleanup should never be needed, but if called on the resulting
+  # 'clear_magic' state it must be a no-op.
+  # ===========================================================================
+  describe '#next_clean_up_step with skip_last_kill' do
+    it 'skips kill and goes directly to clear_magic' do
+      gs = build_game_state(skip_last_kill: true)
+
+      gs.next_clean_up_step
+
+      expect(gs.finish_killing?).to be false
+      expect(gs.finish_spell_casting?).to be true
+    end
+
+    it 'force_cleanup is a no-op when skip_last_kill already skipped kill' do
+      gs = build_game_state(skip_last_kill: true)
+
+      gs.next_clean_up_step
+      gs.force_cleanup
+
+      expect(gs.finish_spell_casting?).to be true
+    end
+  end
+
+  # ===========================================================================
+  # Predicate consistency
+  #
+  # Validates that the predicates agree with the state after force_cleanup.
+  # A mismatch here could cause the main loop to get stuck or skip steps.
+  # ===========================================================================
+  describe 'predicate consistency after force_cleanup' do
+    it 'cleaning_up? remains true' do
+      gs = build_game_state(clean_up_step: 'kill')
+
+      gs.force_cleanup
+
+      expect(gs.cleaning_up?).to be true
+    end
+
+    it 'done_cleaning_up? is false' do
+      gs = build_game_state(clean_up_step: 'kill')
+
+      gs.force_cleanup
+
+      expect(gs.done_cleaning_up?).to be false
+    end
+
+    it 'finish_killing? is false' do
+      gs = build_game_state(clean_up_step: 'kill')
+
+      gs.force_cleanup
+
+      expect(gs.finish_killing?).to be false
+    end
+
+    it 'stowing? is false' do
+      gs = build_game_state(clean_up_step: 'kill')
+
+      gs.force_cleanup
+
+      expect(gs.stowing?).to be false
+    end
+  end
+end
+
+# ###################################################################
+# MERGED FROM spec/combat_trainer_almanac_spec.rb
+# ###################################################################
+
+RSpec.describe TrainerProcess do
+  def build_trainer(**overrides)
+    instance = TrainerProcess.allocate
+    defaults = {
+      almanac: 'almanac',
+      almanac_skills: [],
+      almanac_priority_skills: [],
+      equipment_manager: double('EquipmentManager')
+    }
+    defaults.merge(overrides).each do |k, v|
+      instance.instance_variable_set(:"@#{k}", v)
+    end
+    instance
+  end
+
+  def build_game_state(**attrs)
+    defaults = {
+      currently_whirlwinding: false,
+      npcs: []
+    }
+    state = double('GameState', defaults.merge(attrs))
+    allow(state).to receive(:sheath_whirlwind_offhand)
+    allow(state).to receive(:wield_whirlwind_offhand)
+    allow(state).to receive(:engage_slow)
+    state
+  end
+
+  describe '#use_almanac' do
+    before(:each) do
+      allow(DRC).to receive(:retreat)
+      allow(DRC).to receive(:bput).and_return('Roundtime')
+      allow(DRCI).to receive(:get_item_if_not_held?).and_return(true)
+      allow(DRCI).to receive(:in_hands?).and_return(true)
+      allow(DRCI).to receive(:exists?).and_return(true)
+      allow(DRCI).to receive(:put_away_item?).and_return(true)
+      UserVars.almanac_last_use = Time.now - 700
+    end
+
+    # -----------------------------------------------------------------
+    # Early return guards
+    # -----------------------------------------------------------------
+    context 'when @almanac is nil' do
+      it 'returns immediately without any game commands' do
+        trainer = build_trainer(almanac: nil)
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).not_to have_received(:retreat)
+        expect(DRCI).not_to have_received(:get_item_if_not_held?)
+      end
+    end
+
+    context 'when cooldown has not elapsed' do
+      it 'returns immediately without any game commands' do
+        trainer = build_trainer
+        game_state = build_game_state
+        UserVars.almanac_last_use = Time.now
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).not_to have_received(:retreat)
+        expect(DRCI).not_to have_received(:get_item_if_not_held?)
+      end
+    end
+
+    context 'when left hand is full and not whirlwinding' do
+      it 'returns immediately' do
+        $left_hand = 'sword'
+        trainer = build_trainer
+        game_state = build_game_state(currently_whirlwinding: false)
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).not_to have_received(:retreat)
+      end
+    end
+
+    # -----------------------------------------------------------------
+    # Almanac script delegation
+    # -----------------------------------------------------------------
+    context 'when almanac script is running' do
+      before(:each) do
+        allow(Script).to receive(:running?).with('almanac').and_return(true)
+      end
+
+      it 'delegates to $ALMANAC.use_almanac' do
+        almanac_script = double('AlmanacScript')
+        $ALMANAC = almanac_script
+        allow(almanac_script).to receive(:use_almanac).and_return(:ok)
+
+        trainer = build_trainer
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(almanac_script).to have_received(:use_almanac)
+        expect(DRCI).not_to have_received(:get_item_if_not_held?)
+      end
+
+      it 'disables almanac when script returns :not_found' do
+        almanac_script = double('AlmanacScript')
+        $ALMANAC = almanac_script
+        allow(almanac_script).to receive(:use_almanac).and_return(:not_found)
+
+        trainer = build_trainer
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(trainer.instance_variable_get(:@almanac)).to be_nil
+      end
+
+      it 're-wields whirlwind offhand after delegation' do
+        almanac_script = double('AlmanacScript')
+        $ALMANAC = almanac_script
+        allow(almanac_script).to receive(:use_almanac).and_return(:ok)
+
+        trainer = build_trainer
+        game_state = build_game_state(currently_whirlwinding: true)
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(game_state).to have_received(:wield_whirlwind_offhand)
+      end
+    end
+
+    # -----------------------------------------------------------------
+    # Successful almanac usage (inline, no almanac script)
+    # -----------------------------------------------------------------
+    context 'when almanac is retrieved successfully' do
+      before(:each) do
+        allow(Script).to receive(:running?).with('almanac').and_return(false)
+      end
+
+      it 'retreats and engages slow before getting the almanac' do
+        trainer = build_trainer
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).to have_received(:retreat).ordered
+        expect(game_state).to have_received(:engage_slow).ordered
+      end
+
+      it 'studies the almanac and puts it away' do
+        trainer = build_trainer
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).to have_received(:bput).with('study my almanac', anything, anything, anything)
+        expect(DRCI).to have_received(:put_away_item?).with('almanac')
+      end
+
+      it 'updates the cooldown timer' do
+        trainer = build_trainer
+        game_state = build_game_state
+        before_time = Time.now
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(UserVars.almanac_last_use).to be >= before_time
+      end
+
+      it 'does not turn the almanac when no training_skill is set' do
+        trainer = build_trainer
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).not_to have_received(:bput).with(/^turn almanac/, anything, anything)
+      end
+
+      it 'turns the almanac to the training skill when almanac_skills are configured' do
+        allow(DRSkill).to receive(:getxp).and_return(5)
+        allow(DRSkill).to receive(:getrank).and_return(100)
+
+        trainer = build_trainer(almanac_skills: ['Scholarship'])
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).to have_received(:bput).with('turn almanac to Scholarship', 'You turn', 'You attempt to turn')
+      end
+
+      it 'prefers priority skills over regular almanac skills' do
+        allow(DRSkill).to receive(:getxp).and_return(5)
+        allow(DRSkill).to receive(:getrank).and_return(100)
+
+        trainer = build_trainer(
+          almanac_skills: ['Scholarship'],
+          almanac_priority_skills: ['Tactics']
+        )
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).to have_received(:bput).with('turn almanac to Tactics', 'You turn', 'You attempt to turn')
+      end
+
+      it 're-wields whirlwind offhand after studying' do
+        trainer = build_trainer
+        game_state = build_game_state(currently_whirlwinding: true)
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(game_state).to have_received(:wield_whirlwind_offhand)
+      end
+
+      it 'sheaths whirlwind offhand before getting the almanac' do
+        trainer = build_trainer
+        game_state = build_game_state(currently_whirlwinding: true)
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(game_state).to have_received(:sheath_whirlwind_offhand)
+      end
+    end
+
+    # -----------------------------------------------------------------
+    # Almanac not found -- disables for the hunt
+    # -----------------------------------------------------------------
+    context 'when almanac is not found in inventory' do
+      before(:each) do
+        allow(Script).to receive(:running?).with('almanac').and_return(false)
+        allow(DRCI).to receive(:get_item_if_not_held?).and_return(false)
+        allow(DRCI).to receive(:in_hands?).and_return(false)
+        allow(DRCI).to receive(:exists?).and_return(false)
+      end
+
+      it 'disables almanac usage for the rest of the hunt' do
+        trainer = build_trainer
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(trainer.instance_variable_get(:@almanac)).to be_nil
+      end
+
+      it 'does not attempt to study or stow' do
+        trainer = build_trainer
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).not_to have_received(:bput).with(/study/, anything, anything, anything)
+        expect(DRCI).not_to have_received(:put_away_item?)
+      end
+
+      it 'does not update the cooldown timer' do
+        trainer = build_trainer
+        game_state = build_game_state
+        UserVars.almanac_last_use = Time.now - 700
+        old_time = UserVars.almanac_last_use
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(UserVars.almanac_last_use).to eq(old_time)
+      end
+
+      it 're-wields whirlwind offhand even on failure' do
+        trainer = build_trainer
+        game_state = build_game_state(currently_whirlwinding: true)
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(game_state).to have_received(:wield_whirlwind_offhand)
+      end
+    end
+
+    # -----------------------------------------------------------------
+    # Hands full -- almanac exists but could not be retrieved
+    # -----------------------------------------------------------------
+    context 'when hands are full but almanac exists' do
+      before(:each) do
+        allow(Script).to receive(:running?).with('almanac').and_return(false)
+        allow(DRCI).to receive(:get_item_if_not_held?).and_return(false)
+        allow(DRCI).to receive(:in_hands?).and_return(false)
+        allow(DRCI).to receive(:exists?).and_return(true)
+      end
+
+      it 'does not disable almanac usage' do
+        trainer = build_trainer
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(trainer.instance_variable_get(:@almanac)).to eq('almanac')
+      end
+
+      it 'returns without studying or stowing' do
+        trainer = build_trainer
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).not_to have_received(:bput).with(/study/, anything, anything, anything)
+        expect(DRCI).not_to have_received(:put_away_item?)
+      end
+    end
+
+    # -----------------------------------------------------------------
+    # Skill selection edge cases
+    # -----------------------------------------------------------------
+    context 'when all almanac_skills are at mindstate 18+' do
+      before(:each) do
+        allow(Script).to receive(:running?).with('almanac').and_return(false)
+        allow(DRSkill).to receive(:getxp).and_return(18)
+        allow(DRSkill).to receive(:getrank).and_return(100)
+      end
+
+      it 'falls back to skill_with_lowest_mindstate' do
+        skill_data = double('SkillData', name: 'Forging', exp: 1, rank: 50)
+        allow(DRSkill).to receive(:list).and_return([skill_data])
+
+        trainer = build_trainer(almanac_skills: ['Scholarship'])
+        game_state = build_game_state
+
+        trainer.send(:use_almanac, game_state)
+
+        expect(DRC).to have_received(:bput).with('turn almanac to Forging', 'You turn', 'You attempt to turn')
+      end
+    end
   end
 end
